Use a unique header file name
The Arduino IDE comes with a TFT library which contains a file named TFT.h. When multiple libraries are found that match an #include directive, the Arduino IDE gives preference to the one whose folder name matches the header file name. For this reason, it was previously impossible to use this library without first deleting the official Arduino TFT library.

To solve this issue, a header file with a unique name was added to the library, and the example sketches updated to use that file name in their #include directives.